Round1A/requirements.txt
pymupdf==1.23.7
huggingface_hub
numpy
Pillow
doclayout_yolo
